Last refactoring before project externalisation.
Imported from ariatemplates/editors-tools (https://github.com/ariatemplates/editors-tools/tree/ultimate-poc).

Makes the old folder the new root of this project, filtering the existence of upper components in history.

Also changed the package definition to start with a new fresh semantic versioning, and a consistent package name (same as the one of the repository).