The HTML parser is now far more permissive - even if it still can't parse successfully any input.
- added some standard colors (for easy stylesheet definition)
- changed the HTML stylesheet for easier visual recognition of elements

Made it more permissive and stores errors in nodes (flag them as having errors and set a property with a single error or a list). Also fixed some issues by the way.

Also, some nodes are marked as possible orphans: we should be able to parse them as they are, but outside of the context of the parent they should be in. In practice, rules have not been updated and these nodes can't be real orphans yet.

Fixed:

- directive content definition
- cdata content is now optional
- unquoted attribute values are now better parsed (alpha)

Elements that can be left open:

- inline
- cdata
- block
- strings
- comments
- opening tag of blocks
- closing tag of blocks

Other:

- directive: all the elements except opening delimiter (that's how we determine this is the directive which is in error state) are now optional.
- attribute: attribute to which a value is expected to be assigned - that is where the assignement operator is present - can be left without any value