Now CodeMirror can be used along with Ace to edit the document.
The content is in sync in the two editors, even if there is an issue with CodeMirror: you need to make a change to the document content once to see the actual content (if it has been edited by Ace just before). The content is synchronized when the editor tab are opened (clicked).

This changed a little bit the architecture of methods, to handle the two editors properly without duplicate changes or things like this.

GUI library:

- it is now possible to disable a button used for an action
- it is possible to add an event handler for a click on a tab