Fixed the binary definition and the current working directory.
Added a binary script in a dedicated folder. This script has a shebang for more portability/reliability. Changed its name: now it corresponds to the name of the project.

Also fixed the mess with the current working directory on which were based some features. Now the cwd is set to the root of the package.