Used browerify to test HTML parsers in the browser.
The HTML parsers have been processed by browserify, to be able to use them inside browsers. This is mainly for performances profiling.

An attempt is ongoing for the AT-HTML parsers, but there is a problem of transport configuration (I guess) when trying to use the ariatemplates nodejs package, for the validation of the graph.