Addition of the extension information on the parsed resource
This commit introduces the possibility to specify an extension for the resource that has to be parsed. This extension is given as argument to the parser so that it can produce the correct tree according to the file extension. This way, .tpl and .tml files can be parsed with the same grammar.

* AT-HTML grammar
Addition change of the root statement validation according to the file extension.

* Code
When an instance of Code is created, the extension is provided as argument.

* Tests
Addition to tests and possibility to specify the extension of the content to test.