Added benchmarks.
A new utility has been created in tests, in order to run benchmarks.

This utility relies on a benchmarking library, and mainly takes care of getting the inputs and displaying the results.

For now it can use two benchmarking libraries:

- the renowned benchmark.js library
- a custom one developed internally, to have more control on how to benchmark and get significant results

It uses the second by default.

Here are the changes that have been done along with that:

- added npm script definition tun run the benchmarks
- added development dependencies for testing, mainly benchmarking
- added inputs to benchmark the parsers HTML and AT-HTML